requirements.txt
# Backend
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
requests==2.32.3
Werkzeug==3.1.3
python-dotenv==1.0.1

# Pruebas
pytest-xdist==3.6.1
